+++ recipes/profiles.rb
@@ -1,0 +1,6 @@
+#
+# Cookbook:: organization-attributes
+# Recipe:: profiles
+#
+
+organization_profiles 'organization'